moved colorize to pl and changed behavior of imshow to automatically colorize up to 20 channels
