+++ spatial_data/pl/__init__.py
@@ -1,4 +1,4 @@
-from .cmap import _get_linear_colormap, _get_listed_colormap
+from .utils import _get_linear_colormap, _get_listed_colormap
 from .plot import PlotAccessor
 from .spectra import plot_expression_spectra
